refactor following pr suggestions also fixed a naming inconsistency
