+++ ts/features/idpay/wallet/screens/AvailableInitiativesListScreen.tsx
@@ -2,7 +2,7 @@ import * as pot from "@pagopa/ts-commons/lib/pot";
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
 import React from "react";
-import { ScrollView, View } from "react-native";
+import { ScrollView, StyleSheet, View } from "react-native";
 import {
   IOLogoPaymentType,
   LogoPayment
@@ -11,17 +11,18 @@ import { HSpacer, VSpacer } from "../../../../components/core/spacer/Spacer";
 import { H1 } from "../../../../components/core/typography/H1";
 import { H4 } from "../../../../components/core/typography/H4";
 import { IOStyles } from "../../../../components/core/variables/IOStyles";
+import BaseScreenComponent from "../../../../components/screens/BaseScreenComponent";
 import TypedI18n from "../../../../i18n";
 import { IOStackNavigationRouteProps } from "../../../../navigation/params/AppParamsList";
 import { WalletParamsList } from "../../../../navigation/params/WalletParamsList";
 import { useIODispatch, useIOSelector } from "../../../../store/hooks";
+import customVariables from "../../../../theme/variables";
 import { IDPayInitiativesList } from "../components/IDPayInitiativesListComponents";
 import { idPayInitiativesFromInstrumentGet } from "../store/actions";
 import {
   idPayAreInitiativesFromInstrumentErrorSelector,
   idPayInitiativesFromInstrumentSelector
 } from "../store/reducers";
-import { BaseHeader } from "../../../../components/screens/BaseHeader";
 
 export type AvailableInitiativesListScreenNavigationParams = {
   idWallet: string;
@@ -78,12 +79,11 @@ export const IdPayInitiativeListScreen = (props: Props) => {
 
   return (
     <>
-      <BaseHeader
+      <BaseScreenComponent
         headerTitle={TypedI18n.t("idpay.wallet.initiativePairing.navigation")}
         goBack={true}
-      />
-      <ScrollView>
-        <View style={IOStyles.horizontalContentPadding}>
+      >
+        <ScrollView style={styles.container}>
           <H1>{TypedI18n.t("idpay.wallet.initiativePairing.header")}</H1>
           <VSpacer size={16} />
           {maskedPan && (
@@ -94,15 +94,20 @@ export const IdPayInitiativeListScreen = (props: Props) => {
             </View>
           )}
           <VSpacer size={16} />
-        </View>
-        <View style={IOStyles.horizontalContentPadding}>
           <IDPayInitiativesList
             initiatives={idpayInitiatives}
             idWallet={idWallet}
           />
           <VSpacer size={24} />
-        </View>
-      </ScrollView>
+        </ScrollView>
+      </BaseScreenComponent>
     </>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: customVariables.contentPadding
+  }
+});
